Fix typechecking errors
